pyproject.toml
[project]
name = "mas-cyp-con-2024"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.4.0",
    "asyncpg>=0.30.0",
    "bcrypt==4.1.2",
    "beautifulsoup4==4.12.3",
    "certifi==2024.8.30",
    "click==8.1.7",
    "fastcore==1.7.4",
    "fastlite==0.0.11",
    "h11==0.14.0",
    "httpcore==1.0.5",
    "httptools==0.6.1",
    "httpx==0.27.2",
    "idna==3.8",
    "itsdangerous==2.2.0",
    "numpy==2.1.1",
    "oauthlib==3.2.2",
    "packaging==24.1",
    "pandas==2.2.3",
    "passlib==1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic==2.9.0",
    "pydantic-core==2.23.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-fasthtml==0.5.1",
    "python-multipart==0.0.9",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "six==1.16.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlalchemy>=2.0.43",
    "sqlite-minutils==3.37.0.post3",
    "starlette==0.38.4",
    "typing-extensions==4.12.2",
    "tzdata==2024.1",
    "uvicorn==0.30.6",
    "uvloop==0.20.0",
    "watchfiles==0.24.0",
    "websockets==13.0.1",
]
